noice noice noice little fixing needed where the color value should range in float between 0.0 to 1.0 , instead i am using integer, otherwise displaying the color

+++ ImageColor.java
@@ -1,22 +1,22 @@
 public class ImageColor {
-    public float getB() {
+    public int getB() {
         return b;
     }
 
-    public float getG() {
+    public int getG() {
         return g;
     }
 
-    public float getR() {
+    public int getR() {
         return r;
     }
 
-    private float r;
+    private int r;
 
-    private float g;
-    private float b;
+    private int g;
+    private int b;
 
-    public ImageColor(float r, float g, float b){
+    public ImageColor(int r, int g, int b){
         this.r = r;
         this.b = b;
         this.g = g;
